split script into tx speicfic logic and runnner logic
